# Up to 2x faster prodigal