--- a/dprog.c
+++ b/dprog.c
@@ -28,7 +28,7 @@
   coding, RBS scores, etc.
 *******************************************************************************/
 
-int dprog(struct _node *nod, int nn, struct _training *tinf, int flag) {
+int dprog(struct _node *nod, int nn, struct _training *tinf, int flag, int max_node_dist) {
   int i, j, min, max_ndx = -1, path, nxt, tmp;
   double max_sc = -1.0;
 
@@ -42,15 +42,15 @@ int dprog(struct _node *nod, int nn, struct _training *tinf, int flag) {
 
     /* Set up distance constraints for making connections, */
     /* but make exceptions for giant ORFS.                 */
-    if(i < MAX_NODE_DIST) min = 0; else min = i-MAX_NODE_DIST;
+    if(i < max_node_dist) min = 0; else min = i-max_node_dist;
     if(nod[i].strand == -1 && nod[i].type != STOP && nod[min].ndx >=
        nod[i].stop_val)
       while(min >= 0 && nod[i].ndx != nod[i].stop_val) min--;
     if(nod[i].strand == 1 && nod[i].type == STOP && nod[min].ndx >=
        nod[i].stop_val)
       while(min >= 0 && nod[i].ndx != nod[i].stop_val) min--;
-    if(min < MAX_NODE_DIST) min = 0;
-    else min = min-MAX_NODE_DIST;
+//    if(min < max_node_dist) min = 0;
+//    else min = min-max_node_dist;
     for(j = min; j < i; j++) {
       score_connection(nod, j, i, tinf, flag);
     }